Move tensorflow-related utilities tf_utils
